requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.6
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3